refactor: reorganizar middlewares y configurar tests
