+++ backend/src/middlewares/auth.js
@@ -1,15 +1,28 @@
-// Middleware para verificar JWT
 import jwt from 'jsonwebtoken';
 
-export const autenticarJWT = (req, res, next) => {
+// Middleware para verificar JWT y añadir el usuario autenticado a req.user
+export default function authMiddleware(req, res, next) {
   const authHeader = req.headers['authorization'];
+
+  // Formato esperado: "Bearer TOKEN"
   const token = authHeader && authHeader.split(' ')[1];
+
   if (!token) {
-    return res.status(401).json({ error: 'Token no proporcionado' });
+    return res.status(401).json({
+      error: 'Token no proporcionado'
+    });
   }
+
   jwt.verify(token, process.env.JWT_SECRET || 'secret', (err, user) => {
-    if (err) return res.status(403).json({ error: 'Token inválido' });
+    if (err) {
+      return res.status(403).json({
+        error: 'Token inválido'
+      });
+    }
+
+    // Guardamos el usuario en la request
     req.user = user;
+
     next();
   });
-};
+}
